requirements.txt
requests
mysql-connector-python
flask
streamlit
pandas
